authapp/authser/requirements.txt
djangorestframework==3.12.2
djongo==1.3.3
PyJWT==1.6.4
pymongo==3.11.2